dae/oprofiled.c: don't set cpu_number when !--separate=cpu, fixing multiple open of the same sample files on SMP. libop/tests/*.c minor fix pointed by build on ia64.

+++ libop/tests/alloc_counter_tests.c
@@ -130,7 +130,7 @@ static void show_counter_map(size_t const * counter_map, size_t nr_events)
 {
 	size_t i;
 	for (i = 0; i < nr_events; ++i)
-		printf("%d ", counter_map[i]);
+		printf("%lu ", (unsigned long)counter_map[i]);
 	printf("\n");
 }
